Move functions from util to fits_support
