+++ src/stdatamodels/util.py
@@ -89,23 +89,6 @@ def gentle_asarray(a, dtype):
         return a
 
 
-def get_short_doc(schema):
-    title = schema.get('title', None)
-    description = schema.get('description', None)
-    if description is None:
-        description = title or ''
-    else:
-        if title is not None:
-            description = title + '\n\n' + description
-    return description.partition('\n')[0]
-
-
-def ensure_ascii(s):
-    if isinstance(s, bytes):
-        s = s.decode('ascii')
-    return s
-
-
 def create_history_entry(description, software=None):
     """
     Create a HistoryEntry object.
